fix(energy): merge pages
