+++ src/pages/Energy/ui/pages/SelectPackages/images/genius.png
