+++ src/pages/Energy/ui/pages/SelectPackages/images/prodigy.png
