+++ src/pages/Energy/ui/pages/SelectPackages/images/smart.png
