+++ src/router.tsx
@@ -60,7 +60,6 @@ import Settings from './pages/Settings/Settings';
 import GovernanceRoutes from './containers/governance/GovernanceRoutes';
 import StudioWrapper from './features/studio/StudioWrapper';
 import Energy from './pages/Energy/Energy';
-import Buy from './pages/Portal/Buy/Buy';
 
 type WrappedRouterProps = {
   children: React.ReactNode;
@@ -193,8 +192,7 @@ function AppRouter() {
           <Route path="/release" element={<Release />} />
           <Route path="/portal" element={<MainPartal />} />
           <Route path="/portal/map" element={<Map />} />
-          <Route path="/buy" element={<Buy />} />
-
+          <Route path="/buy/*" element={<Energy />} />
           {/* wasm */}
           <Route path="/libs" element={<Codes />} />
           <Route path="/libs/:codeId" element={<CodePage />} />
@@ -216,8 +214,6 @@ function AppRouter() {
           )} */}
 
           <Route path="/keys" element={<Keys />} />
-
-          <Route path="/energy/*" element={<Energy />} />
 
           <Route path="/settings/*" element={<Settings />} />
